% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{data}
\alias{data}
\title{data}
\format{
A data frame with 1860 rows and 4 variables
}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/00Index.html}
}
\description{
Sample data based on dataset EuStockMarkets in the datasets package.
}
\examples{
data(mycsv)
thismodel=lm(formula=DAX ~ .,data=data)
expectedprob=predict(thismodel,data)
actualprob=data$DAX
residual=actualprob-expectedprob
data=cbind(data,expectedprob, actualprob, residual)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{data}
\keyword{sample}
