% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{fileConn}
\alias{fileConn}
\title{fileConn}
\format{A text output object.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(fileConn)
head(fileConn)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{fileConn}
