% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{futuresdata}
\alias{futuresdata}
\title{futuresdata}
\format{A data frame with 1860 rows and 4 variables}
\source{
\url{stat.ethz.ch/R-manual/R-devel/library/datasets/html/00Index.html}
}
\description{
Sample futures data based on dataset EuStockMarkets in the datasets package.
}
\examples{
data(futuresdata)
head(futuresdata)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{futuresdata}
