% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\name{generateResidualCutoffCode}
\alias{generateResidualCutoffCode}
\title{generateResidualCutoffCode}
\usage{
generateResidualCutoffCode(data, filename, trees, names, runname, ...)
}
\arguments{
\item{data}{A dataframe}

\item{filename}{A string}

\item{trees}{A list of trees generated by saveTree}

\item{names}{A list of level names}

\item{runname}{A string corresponding to the name of the factor variable being analysed}

\item{...}{and parameters to be passed through}
}
\value{
A list of residuals for each tree provided.
}
\description{
For each tree print a summary of the significant residuals as specified by the user
}
\keyword{saveTree}
