% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\name{generateTrees}
\alias{generateTrees}
\title{generateTrees}
\usage{
generateTrees(data, vars, expr, runname, ...)
}
\arguments{
\item{data}{A dataframe}

\item{vars}{A list of candidate predictors}

\item{expr}{A expression to be modelled by the RPART tree}

\item{runname}{A string corresponding to the name of the variable being modelled}

\item{...}{and parameters to be passed through}
}
\value{
A list of residual trees for each level of the mainfac factor provided
}
\description{
Generate a residual tree for each level of factor mainfac
}
\keyword{generateTrees}
