% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{names}
\alias{names}
\title{names}
\format{A list of variable names used in examples.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(names)
head(names)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{names}
