% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\name{parseSplits}
\alias{parseSplits}
\title{parseSplits}
\usage{
parseSplits(thistree)
}
\arguments{
\item{thistree}{A tree}
}
\value{
A list of parsed splits.
}
\description{
Extract information relating to the paths and volume of data in the leaves of the tree
}
\keyword{saveTree}
