% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\name{printResiduals}
\alias{printResiduals}
\title{printResiduals}
\usage{
printResiduals(fileConn, all, dat, runname, levelname,
  treeSummaryResidualThreshold, treeSummaryMinBucket,
  treeSummaryResidualMagnitudeThreshold, ...)
}
\arguments{
\item{fileConn}{A file connection}

\item{all}{A dataframe}

\item{dat}{The dataset}

\item{runname}{A string corresponding to the name of the factor being analysed}

\item{levelname}{A string corresponding to the factor level being analysed}

\item{treeSummaryResidualThreshold}{The minimum residual threshold}

\item{treeSummaryMinBucket}{The minumum volume per leaf}

\item{treeSummaryResidualMagnitudeThreshold}{Minimun residual magnitude}

\item{...}{and parameters to be passed through}
}
\value{
Residuals are printed and also saved in a simplified format.
}
\description{
This function generates a residual tree on a subset of the data
}
\examples{

require(featurefinder)
data(examples)
printResiduals(fileConn,splitlist[t][[1]],dat, runname, names[t],
  treeSummaryResidualThreshold,treeSummaryMinBucket,
  treeSummaryResidualMagnitudeThreshold)
}
\keyword{saveTree}
