% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{runname}
\alias{runname}
\title{runname}
\format{A string corresponding to the name of the variable being modelled}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(runname)
head(runname)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{runname}
