% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{splitlist}
\alias{splitlist}
\title{splitlist}
\format{Sample list of node split formulae.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(splitlist)
head(splitlist)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{splitlist}
