% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{treeGenerationMinBucket}
\alias{treeGenerationMinBucket}
\title{treeGenerationMinBucket}
\format{Minimum number of data points per leaf allowed in tree generation.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(treeGenerationMinBucket)
head(treeGenerationMinBucket)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{treeGenerationMinBucket}
