% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{trees}
\alias{trees}
\title{trees}
\format{A data frame with 234 rows and 11 variables}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(trees)
head(trees)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{trees}
