% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{treeSummaryResidualMagnitudeThreshold}
\alias{treeSummaryResidualMagnitudeThreshold}
\title{treeSummaryResidualMagnitudeThreshold}
\format{Minimum allowed residual magnitude in leaf summary generation.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(treeSummaryResidualMagnitudeThreshold)
head(treeSummaryResidualMagnitudeThreshold)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{treeSummaryResidualMagnitudeThreshold}
