% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{treeSummaryResidualThreshold}
\alias{treeSummaryResidualThreshold}
\title{treeSummaryResidualThreshold}
\format{Minimum allowed residual value in leaf summary generation.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(treeSummaryResidualThreshold)
head(treeSummaryResidualThreshold)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{treeSummaryResidualThreshold}
