% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{mainfaclevels}
\alias{mainfaclevels}
\title{mainfaclevels}
\format{Levels of the main or current factor being scanned.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(mainfaclevels)
head(mainfaclevels)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{mainfaclevels}
