% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{maxFactorLevels}
\alias{maxFactorLevels}
\title{maxFactorLevels}
\format{Maximum allowable number of factor levels before the variable is converted to numeric.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(maxFactorLevels)
head(maxFactorLevels)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{maxFactorLevels}
