% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurefinder.r
\docType{data}
\name{treesAll}
\alias{treesAll}
\title{treesAll}
\format{Full dataset tree example.}
\source{
\url{ggplot2.org}
}
\description{
Sample data based on dataset mpg in the ggplot2 package
}
\examples{
data(treesAll)
head(treesAll)
}
\author{
Richard Davis \email{richard.davis@cba.com.au}
}
\keyword{treesAll}
